requirements.txt
feedparser>=6.0
anthropic>=0.40.0
jinja2>=3.1
python-dotenv>=1.0
requests>=2.31
